chrome-transcriber-backend/requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
moviepy
